+++ src/main/java/com/telepathicgrunt/repurposedstructures/mixin/structures/StructureMobSpawningMixin.java
@@ -1,6 +1,6 @@
 package com.telepathicgrunt.repurposedstructures.mixin.structures;
 
-import com.telepathicgrunt.repurposedstructures.misc.StructureMobSpawning;
+import com.telepathicgrunt.repurposedstructures.misc.MobSpawningOverTime;
 import net.minecraft.entity.SpawnGroup;
 import net.minecraft.util.collection.Pool;
 import net.minecraft.util.math.BlockPos;
@@ -22,7 +22,7 @@ public class StructureMobSpawningMixin {
             cancellable = true
     )
     private void repurposedstructures_structureMobs(Biome biome, StructureAccessor accessor, SpawnGroup group, BlockPos pos, CallbackInfoReturnable<Pool<SpawnSettings.SpawnEntry>> cir) {
-        Pool<SpawnSettings.SpawnEntry> pool = StructureMobSpawning.getStructureSpawns(biome, accessor, group, pos);
+        Pool<SpawnSettings.SpawnEntry> pool = MobSpawningOverTime.getStructureSpawns(biome, accessor, group, pos);
         if(pool != null) cir.setReturnValue(pool);
     }
 }